feat : 게시물 타입 지정
