--- src/types/User.ts
@@ -1,6 +1,0 @@
-export interface User {
-  email: string;
-  password: string;
-  nickname: string;
-  name: string;
-}
